# openapi.yaml
openapi: 3.0.0
paths:
...
  /auth/google:
    get:
      operationId: AuthController_googleLogin
      parameters: []
      responses:
        '200':
          description: ''
        '302':
          description: Redireciona para o Google OAuth
      summary: Inicia o fluxo OAuth Google para membros
      tags: &ref_0
        - auth
  /auth/google/callback:
    get:
      operationId: AuthController_googleCallback
      parameters:
        - name: invite
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Login bem-sucedido, retorna tokens de acesso
        '401':
          description: Autenticação falhou
      summary: Callback do OAuth Google — retorna JWT e dados do usuário
      tags: *ref_0
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Novos tokens gerados com sucesso
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '401':
          description: Refresh token inválido ou revogado
      security:
        - bearer: []
      summary: Renova o access token usando o refresh token
      tags: *ref_0
  /auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Sessão encerrada com sucesso
        '201':
          description: ''
      security:
        - bearer: []
      summary: Revoga o refresh token e encerra a sessão
      tags: *ref_0
  /auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        '200':
          description: Dados do usuário
        '401':
          description: Não autenticado
      security:
        - bearer: []
      summary: Retorna os dados do usuário autenticado
      tags: *ref_0
...
  /admin/auth/google:
    get:
      operationId: AdminAuthController_googleLogin
      parameters: []
      responses:
        '200':
          description: ''
        '302':
          description: Redireciona para o Google OAuth (admin)
      summary: Inicia o fluxo OAuth Google para administradores
      tags: &ref_1
        - admin/auth
  /admin/auth/google/callback:
    get:
      operationId: AdminAuthController_googleCallback
      parameters: []
      responses:
        '200':
          description: Login admin bem-sucedido
        '403':
          description: Usuário não possui role admin
      summary: Callback OAuth Google (admin) — valida role e retorna JWT
      tags: *ref_1
  /admin/auth/refresh:
    post:
      operationId: AdminAuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Novos tokens gerados
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '401':
          description: Refresh token inválido
      security:
        - bearer: []
      summary: Renova o access token (admin)
      tags: *ref_1
  /admin/auth/logout:
    post:
      operationId: AdminAuthController_logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Sessão encerrada
        '201':
          description: ''
      security:
        - bearer: []
      summary: Encerra a sessão do admin
      tags: *ref_1
  /admin/users:
    get:
      operationId: AdminUsersController_listUsers
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: search
          required: false
          in: query
          schema:
            type: string
        - name: role
          required: false
          in: query
          schema:
            enum:
              - member
              - admin
            type: string
        - name: isActive
          required: false
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista paginada de membros
      security: &ref_2
        - bearer: []
      summary: Lista membros com paginação e filtros
      tags: &ref_3
        - admin/users
  /admin/users/{id}:
    get:
      operationId: AdminUsersController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Dados do membro
        '404':
          description: Membro não encontrado
      security: *ref_2
      summary: Retorna os detalhes de um membro específico
      tags: *ref_3
    patch:
      operationId: AdminUsersController_updateUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Membro atualizado
        '404':
          description: Membro não encontrado
      security: *ref_2
      summary: Atualiza nome, roles ou status ativo de um membro
      tags: *ref_3
    delete:
      operationId: AdminUsersController_softDeleteUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Membro desativado (soft delete)
        '404':
          description: Membro não encontrado
      security: *ref_2
      summary: Soft delete de um membro
      tags: *ref_3
  /admin/users/{id}/points:
    patch:
      operationId: AdminUsersController_adjustPoints
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustPointsDto'
      responses:
        '200':
          description: Pontos ajustados, transação criada
        '404':
          description: Membro não encontrado
      security: *ref_2
      summary: Ajuste manual de pontos de um membro
      tags: *ref_3
...
  /points:
    get:
      operationId: PointsController_getSummary
      parameters: []
      responses:
        '200':
          description: Resumo de pontos e gamificação
      security: &ref_8
        - bearer: []
      summary: >-
        Retorna o resumo de gamificação do membro (saldo, streak, nível, total
        ganho)
      tags: &ref_9
        - Pontos
  /points/history:
    get:
      operationId: PointsController_getHistory
      parameters:
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
      responses:
        '200':
          description: Histórico paginado de transações
      security: *ref_8
      summary: Retorna o histórico paginado de transações de pontos do membro
      tags: *ref_9
...
  /tasks/{id}:
    get:
      operationId: TasksController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Detalhes da tarefa
        '404':
          description: Tarefa não encontrada
      security: *ref_18
      summary: Retorna o detalhe de uma tarefa do membro
      tags: *ref_19
...
  /admin/task-instances:
    get:
      operationId: AdminTaskInstancesController_findAll
      parameters:
        - name: date
          required: false
          in: query
          schema:
            example: '2026-07-01'
            type: string
        - name: status
          required: false
          in: query
          schema:
            enum:
              - pending
              - in_progress
              - done
              - overdue
              - skipped
            type: string
        - name: userId
          required: false
          in: query
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
      responses:
        '200':
          description: Lista de instâncias de tarefas
      security: &ref_20
        - bearer: []
      summary: Lista instâncias de tarefas com filtros opcionais
      tags: &ref_21
        - admin/task-instances
  /admin/task-instances/{id}:
    get:
      operationId: AdminTaskInstancesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Detalhes da instância
        '404':
          description: Instância não encontrada
      security: *ref_20
      summary: Retorna o detalhe de uma instância de tarefa
      tags: *ref_21
    patch:
      operationId: AdminTaskInstancesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskInstanceDto'
      responses:
        '200':
          description: Instância atualizada
      security: *ref_20
      summary: Edita campos de override de uma ocorrência avulsa (is_exception = true)
      tags: *ref_21
    delete:
      operationId: AdminTaskInstancesController_softDelete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Instância removida
      security: *ref_20
      summary: Soft delete de uma instância de tarefa
      tags: *ref_21
info:
  title: domy-api
  description: API do app doméstico gamificado
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 10
      required:
        - refreshToken
    AuthUserDto:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
            enum:
              - member
              - admin
          description: Roles do usuário
        level:
          enum:
            - Bronze
            - Prata
            - Ouro
            - Diamante
          type: string
          description: Nível atual do membro
        id:
          type: string
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        pointsBalance:
          type: number
        currentStreak:
          type: number
        longestStreak:
          type: number
        isActive:
          type: boolean
      required:
        - roles
        - level
        - id
        - email
        - name
        - avatarUrl
        - pointsBalance
        - currentStreak
        - longestStreak
        - isActive
    AuthResponseDto:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/AuthUserDto'
        accessToken:
          type: string
        refreshToken:
          type: string
      required:
        - user
        - accessToken
        - refreshToken
    AdjustPointsDto:
      type: object
      properties:
        amount:
          type: number
        description:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - amount
        - description
    UpdateUserDto:
      type: object
      properties:
        roles:
          minItems: 1
          type: array
          items:
            type: string
            enum:
              - member
              - admin
          description: Roles do usuário
        name:
          type: string
          minLength: 1
        isActive:
          type: boolean
...
    UpdateTaskInstanceDto:
      type: object
      properties:
        overrideTitle:
          type: string
          maxLength: 255
          description: Título sobreposto para esta ocorrência
        overrideDescription:
          type: string
          description: Descrição sobreposta para esta ocorrência
        overrideDeadlineAt:
          type: string
          description: Prazo sobreposto para esta ocorrência (ISO 8601)
          example: '2026-07-01T23:59:00.000Z'